backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
lxml>=4.9.0
openai-agents>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
litellm
rdkit
